# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NXTTRACK SaaS Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type TenantPlan = (typeof TenantPlan)[keyof typeof TenantPlan];

export const TenantPlan = {
  free: "free",
  starter: "starter",
  pro: "pro",
  enterprise: "enterprise",
} as const;

export interface Tenant {
  id: string;
  slug: string;
  name: string;
  plan: TenantPlan;
  logoUrl?: string | null;
  primaryColor?: string | null;
  createdAt: string;
  updatedAt: string;
  userCount: number;
}

export type CreateTenantBodyPlan =
  (typeof CreateTenantBodyPlan)[keyof typeof CreateTenantBodyPlan];

export const CreateTenantBodyPlan = {
  free: "free",
  starter: "starter",
  pro: "pro",
  enterprise: "enterprise",
} as const;

export interface CreateTenantBody {
  slug: string;
  name: string;
  plan: CreateTenantBodyPlan;
  logoUrl?: string | null;
  primaryColor?: string | null;
}

export type UpdateTenantBodyPlan =
  (typeof UpdateTenantBodyPlan)[keyof typeof UpdateTenantBodyPlan];

export const UpdateTenantBodyPlan = {
  free: "free",
  starter: "starter",
  pro: "pro",
  enterprise: "enterprise",
} as const;

export interface UpdateTenantBody {
  name?: string;
  plan?: UpdateTenantBodyPlan;
  logoUrl?: string | null;
  primaryColor?: string | null;
}

export type TenantUserRole =
  (typeof TenantUserRole)[keyof typeof TenantUserRole];

export const TenantUserRole = {
  owner: "owner",
  admin: "admin",
  member: "member",
  viewer: "viewer",
} as const;

export interface TenantUser {
  id: string;
  tenantId: string;
  email: string;
  name: string;
  role: TenantUserRole;
  avatarUrl?: string | null;
  createdAt: string;
}

export type CreateTenantUserBodyRole =
  (typeof CreateTenantUserBodyRole)[keyof typeof CreateTenantUserBodyRole];

export const CreateTenantUserBodyRole = {
  owner: "owner",
  admin: "admin",
  member: "member",
  viewer: "viewer",
} as const;

export interface CreateTenantUserBody {
  email: string;
  name: string;
  role: CreateTenantUserBodyRole;
  avatarUrl?: string | null;
}

export interface ActivityEvent {
  id: string;
  tenantId: string;
  userId: string;
  userName: string;
  userAvatarUrl?: string | null;
  action: string;
  resource: string;
  resourceId?: string | null;
  createdAt: string;
}

export type DashboardSummaryMembersByRole = { [key: string]: number };

export interface DashboardSummary {
  totalUsers: number;
  activeUsers: number;
  plan: string;
  storageUsedMb: number;
  storageCapacityMb: number;
  recentActivityCount: number;
  membersByRole: DashboardSummaryMembersByRole;
}

export type GetTenantActivityParams = {
  limit?: number;
};