# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NXTTRACK SaaS Platform API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: tenants
    description: Tenant management
  - name: users
    description: User management
  - name: activity
    description: Activity feed
  - name: dashboard
    description: Dashboard summaries
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /tenants:
    get:
      operationId: listTenants
      tags: [tenants]
      summary: List all tenants
      responses:
        "200":
          description: List of tenants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tenant"
    post:
      operationId: createTenant
      tags: [tenants]
      summary: Create a new tenant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTenantBody"
      responses:
        "201":
          description: Tenant created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"

  /tenants/{slug}:
    get:
      operationId: getTenant
      tags: [tenants]
      summary: Get tenant by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Tenant details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        "404":
          description: Tenant not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateTenant
      tags: [tenants]
      summary: Update a tenant
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTenantBody"
      responses:
        "200":
          description: Tenant updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
    delete:
      operationId: deleteTenant
      tags: [tenants]
      summary: Delete a tenant
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Tenant deleted

  /tenants/{slug}/users:
    get:
      operationId: listTenantUsers
      tags: [users]
      summary: List users for a tenant
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TenantUser"
    post:
      operationId: createTenantUser
      tags: [users]
      summary: Add a user to a tenant
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTenantUserBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantUser"

  /tenants/{slug}/users/{userId}:
    delete:
      operationId: deleteTenantUser
      tags: [users]
      summary: Remove user from tenant
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: User removed

  /tenants/{slug}/activity:
    get:
      operationId: getTenantActivity
      tags: [activity]
      summary: Get recent activity feed for a tenant
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Activity feed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityEvent"

  /tenants/{slug}/dashboard:
    get:
      operationId: getTenantDashboard
      tags: [dashboard]
      summary: Get dashboard summary stats for a tenant
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    Tenant:
      type: object
      properties:
        id:
          type: string
        slug:
          type: string
        name:
          type: string
        plan:
          type: string
          enum: [free, starter, pro, enterprise]
        logoUrl:
          type: string
          nullable: true
        primaryColor:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        userCount:
          type: integer
      required:
        - id
        - slug
        - name
        - plan
        - createdAt
        - updatedAt
        - userCount

    CreateTenantBody:
      type: object
      properties:
        slug:
          type: string
        name:
          type: string
        plan:
          type: string
          enum: [free, starter, pro, enterprise]
        logoUrl:
          type: string
          nullable: true
        primaryColor:
          type: string
          nullable: true
      required:
        - slug
        - name
        - plan

    UpdateTenantBody:
      type: object
      properties:
        name:
          type: string
        plan:
          type: string
          enum: [free, starter, pro, enterprise]
        logoUrl:
          type: string
          nullable: true
        primaryColor:
          type: string
          nullable: true

    TenantUser:
      type: object
      properties:
        id:
          type: string
        tenantId:
          type: string
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [owner, admin, member, viewer]
        avatarUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - tenantId
        - email
        - name
        - role
        - createdAt

    CreateTenantUserBody:
      type: object
      properties:
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [owner, admin, member, viewer]
        avatarUrl:
          type: string
          nullable: true
      required:
        - email
        - name
        - role

    ActivityEvent:
      type: object
      properties:
        id:
          type: string
        tenantId:
          type: string
        userId:
          type: string
        userName:
          type: string
        userAvatarUrl:
          type: string
          nullable: true
        action:
          type: string
        resource:
          type: string
        resourceId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - tenantId
        - userId
        - userName
        - action
        - resource
        - createdAt

    DashboardSummary:
      type: object
      properties:
        totalUsers:
          type: integer
        activeUsers:
          type: integer
        plan:
          type: string
        storageUsedMb:
          type: number
        storageCapacityMb:
          type: number
        recentActivityCount:
          type: integer
        membersByRole:
          type: object
          additionalProperties:
            type: integer
      required:
        - totalUsers
        - activeUsers
        - plan
        - storageUsedMb
        - storageCapacityMb
        - recentActivityCount
        - membersByRole