# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NXTTRACK SaaS Platform API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all tenants
 */
export const ListTenantsResponseItem = zod.object({
  id: zod.string(),
  slug: zod.string(),
  name: zod.string(),
  plan: zod.enum(["free", "starter", "pro", "enterprise"]),
  logoUrl: zod.string().nullish(),
  primaryColor: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
  userCount: zod.number(),
});
export const ListTenantsResponse = zod.array(ListTenantsResponseItem);

/**
 * @summary Create a new tenant
 */
export const CreateTenantBody = zod.object({
  slug: zod.string(),
  name: zod.string(),
  plan: zod.enum(["free", "starter", "pro", "enterprise"]),
  logoUrl: zod.string().nullish(),
  primaryColor: zod.string().nullish(),
});

/**
 * @summary Get tenant by slug
 */
export const GetTenantParams = zod.object({
  slug: zod.coerce.string(),
});

export const GetTenantResponse = zod.object({
  id: zod.string(),
  slug: zod.string(),
  name: zod.string(),
  plan: zod.enum(["free", "starter", "pro", "enterprise"]),
  logoUrl: zod.string().nullish(),
  primaryColor: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
  userCount: zod.number(),
});

/**
 * @summary Update a tenant
 */
export const UpdateTenantParams = zod.object({
  slug: zod.coerce.string(),
});

export const UpdateTenantBody = zod.object({
  name: zod.string().optional(),
  plan: zod.enum(["free", "starter", "pro", "enterprise"]).optional(),
  logoUrl: zod.string().nullish(),
  primaryColor: zod.string().nullish(),
});

export const UpdateTenantResponse = zod.object({
  id: zod.string(),
  slug: zod.string(),
  name: zod.string(),
  plan: zod.enum(["free", "starter", "pro", "enterprise"]),
  logoUrl: zod.string().nullish(),
  primaryColor: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
  userCount: zod.number(),
});

/**
 * @summary Delete a tenant
 */
export const DeleteTenantParams = zod.object({
  slug: zod.coerce.string(),
});

/**
 * @summary List users for a tenant
 */
export const ListTenantUsersParams = zod.object({
  slug: zod.coerce.string(),
});

export const ListTenantUsersResponseItem = zod.object({
  id: zod.string(),
  tenantId: zod.string(),
  email: zod.string(),
  name: zod.string(),
  role: zod.enum(["owner", "admin", "member", "viewer"]),
  avatarUrl: zod.string().nullish(),
  createdAt: zod.coerce.date(),
});
export const ListTenantUsersResponse = zod.array(ListTenantUsersResponseItem);

/**
 * @summary Add a user to a tenant
 */
export const CreateTenantUserParams = zod.object({
  slug: zod.coerce.string(),
});

export const CreateTenantUserBody = zod.object({
  email: zod.string(),
  name: zod.string(),
  role: zod.enum(["owner", "admin", "member", "viewer"]),
  avatarUrl: zod.string().nullish(),
});

/**
 * @summary Remove user from tenant
 */
export const DeleteTenantUserParams = zod.object({
  slug: zod.coerce.string(),
  userId: zod.coerce.string(),
});

/**
 * @summary Get recent activity feed for a tenant
 */
export const GetTenantActivityParams = zod.object({
  slug: zod.coerce.string(),
});

export const getTenantActivityQueryLimitDefault = 20;

export const GetTenantActivityQueryParams = zod.object({
  limit: zod.coerce.number().default(getTenantActivityQueryLimitDefault),
});

export const GetTenantActivityResponseItem = zod.object({
  id: zod.string(),
  tenantId: zod.string(),
  userId: zod.string(),
  userName: zod.string(),
  userAvatarUrl: zod.string().nullish(),
  action: zod.string(),
  resource: zod.string(),
  resourceId: zod.string().nullish(),
  createdAt: zod.coerce.date(),
});
export const GetTenantActivityResponse = zod.array(
  GetTenantActivityResponseItem,
);

/**
 * @summary Get dashboard summary stats for a tenant
 */
export const GetTenantDashboardParams = zod.object({
  slug: zod.coerce.string(),
});

export const GetTenantDashboardResponse = zod.object({
  totalUsers: zod.number(),
  activeUsers: zod.number(),
  plan: zod.string(),
  storageUsedMb: zod.number(),
  storageCapacityMb: zod.number(),
  recentActivityCount: zod.number(),
  membersByRole: zod.record(zod.string(), zod.number()),
});